Create max_sum_of_array.java
There is a straight line of students of various heights. The students' heights are given in the form of an array, in the order they are standing in the line.
Consider the region of a student as the length of the largest subarray that includes that student's position, and in which that student's height is equal to maximum height among all students present in that subarray. Return the sum of the region of all students.
For example-
• heights = [3, 5, 6]
• The longest subarray in which the first student's height is equal to the maximum height among all other students is [3]; thus, the length of the region of the first student is 1.
• The longest subarray in which the second student's height is equal to maximum height among all other students is [3, 5]; thus, the length of the region of the second student is 2.
• The longest subarray in which the third student's height is equal to the maximum height among all other students is [3, 5, 6]; thus, the length of the region of the third student is 3.
• Thus, the sum of the lengths of all regions ons of all students is 1+2+3=6.
Function Description SCOITOIOR
Complete the function calculate TotalRegion in the editor below. The function must return the desired sum of all regions.
calculate Total Region has the following parameter(s): heights, an array of the heights of students standing in the line                                 Sample Input For Custom Testing
3
1
2
1
Sample Output
5
Explanation
Subarrays for values 1, 2, and 1 are [1], [1, 2, 1], and [1]. The sum of their lengths is 1+3+1 = 5.
▼ Sample Case 1
Sample Input For Custom Testing
4
1
1
1
1
Sample Output
16
Explanation For example:
heights [1,1,1,
All elements are the same, so each subarray contains 4 elements. 4 + 4+4+4=16.